Initial upload of all the files to this eclipse plugin.
The following things works already:
The plugin can be started and log4j is correctly set up.
If you press Strg + Space while a Java Editor is open, you get a sample
completion proposal with the name full property and some description
which is nicely displayed.

You can also invoke this completion proposal, and then it replaces the
whole document with an arbitrarily chosen string of mine.

Important for future for is this link
http://grepcode.com/file/repository.grepcode.com/java/eclipse.org/4.3/org.eclipse/text/3.5.300/org/eclipse/jface/text/link/LinkedPositionGroup.java#LinkedPositionGroup.addPosition%28org.eclipse.jface.text.link.LinkedPosition%29

and this link

http://grepcode.com/file/repository.grepcode.com/java/eclipse.org/4.3/org.eclipse.jface/text/3.8.100/org/eclipse/jface/text/templates/TemplateProposal.java#TemplateProposal.apply%28org.eclipse.jface.text.ITextViewer%2Cchar%2Cint%2Cint%29